don't inject rest client into pojos, instead call rest client from DomainCheckAction. separate domain classes (pojos) into separate package
