--- whois-update/src/main/java/net/ripe/db/whois/update/dns/zonemaster/ZonemasterRequest.java
@@ -1,5 +1,0 @@
-package net.ripe.db.whois.update.dns.zonemaster;
-
-interface ZonemasterRequest {
-    Request getRequest();
-}
