+++ whois-update/src/main/java/net/ripe/db/whois/update/dns/zonemaster/domain/GetTestResultsRequest.java
@@ -1,15 +1,15 @@
-package net.ripe.db.whois.update.dns.zonemaster;
+package net.ripe.db.whois.update.dns.zonemaster.domain;
 
 /**
  * Taken from Zonemaster documentation
  *
  * @see <a href="https://github.com/dotse/zonemaster-backend/blob/master/docs/API.md">Zonemaster documentation</a>
  */
-class GetTestResultsRequest extends ZonemasterRequestSupport {
+public class GetTestResultsRequest implements ZonemasterRequest {
 
     final Request request;
 
-    GetTestResultsRequest(String id) {
+    public GetTestResultsRequest(String id) {
         this.request = new Request();
         request.setMethod(Request.Method.GET_TEST_RESULTS);
         final Request.Params params = new Request.Params();
@@ -20,10 +20,5 @@ class GetTestResultsRequest extends ZonemasterRequestSupport {
     @Override
     public Request getRequest() {
         return request;
-    }
-
-    public static void main(String[] args) {
-        GetTestResultsRequest req = new GetTestResultsRequest("i ij i ij oo");
-        System.out.println("request:\n" + req.getRequest());
     }
 }
