+++ whois-update/src/main/java/net/ripe/db/whois/update/dns/zonemaster/domain/TestProgressRequest.java
@@ -1,15 +1,15 @@
-package net.ripe.db.whois.update.dns.zonemaster;
+package net.ripe.db.whois.update.dns.zonemaster.domain;
 
 /**
  * Taken from Zonemaster documentation
  *
  * @see <a href="https://github.com/dotse/zonemaster-backend/blob/master/docs/API.md">Zonemaster documentation</a>
  */
-class TestProgressRequest extends ZonemasterRequestSupport {
+public class TestProgressRequest implements ZonemasterRequest {
 
     final Request request;
 
-    TestProgressRequest(final String id) {
+    public TestProgressRequest(final String id) {
         this.request = new Request();
         this.request.setMethod(Request.Method.TEST_PROGRESS);
         final Request.Params params = new Request.Params();
