+++ whois-update/src/main/java/net/ripe/db/whois/update/dns/zonemaster/domain/ZonemasterRequest.java
@@ -1,0 +1,6 @@
+package net.ripe.db.whois.update.dns.zonemaster.domain;
+
+// TODO: remove getRequest(), just implement interface
+public interface ZonemasterRequest {
+    Request getRequest();
+}
